% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregress.R
\docType{package}
\name{aggregress}
\alias{aggregress}
\title{aggregress: A package for regression analysis with aggregated data.}
\description{
The aggregress package makes adjustments to the lm object output
when the inputted data are aggregated.
The resulting object is indistinguishable from that from the
original unaggregated data.
It includes adjustments to the statistics and diagnostics.
}
