% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregress.R
\name{gen_agg_lm}
\alias{gen_agg_lm}
\title{Generate data to test the aggregress package.}
\usage{
gen_agg_lm(model)
}
\arguments{
\item{model}{a string to denote the model with which
to generate the data frame. It can be 'lpm' for a linear
proability model, 'lm' for a linear regression model
or 'log' for a logistic regression model.}
}
\value{
a data frame of response variables and covariates for
linear regression.
}
\description{
\code{gen_agg_lm} generates data to test the \code{aggregress}
package. It is an internal function used for testing.
}
