% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregress.R
\name{summary_agg_lm}
\alias{summary_agg_lm}
\title{Summarizing Linear Model Fits with Aggregated Data}
\usage{
summary_agg_lm(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{an object of class "\code{agg_lm}",
usually a result of a call to \code{agg_lm}.}
}
\value{
\code{summary_agg_lm} computes and returns a list of summary statistics
of the fitted linear model given in \code{object}.
}
\description{
\code{summary} method for class \code{agg_lm}.
}
\seealso{
\code{summary.lm} function in the \code{stats} library.
\code{agg_lm} for model fits to produce objects in class "\code{agg_lm}".
}
