% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregress_het.R
\name{white_hccme}
\alias{white_hccme}
\title{Calculate the Heteroskedasticity-Consistent Covariance Matrix Estimator.}
\usage{
white_hccme(lm_in)
}
\arguments{
\item{lm_in}{is an object of class "\code{lm}" that is the output of a
linear regression model.}
}
\value{
A data frame of the form of \code{coef(summary(lm_in))},
except with standard errors and the remaining statistics adjusted
for heteroskedasticity.
}
\description{
\code{white_hccme} is a helper function to calculate the HCCME,
to test the calculation for un-aggregated data, i.e. data
with weights either \code{NULL} or equal to 1.
}
\seealso{
The \code{vcovHC} function in the \code{sandwich} package.
}
