% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregress_het.R
\name{white_hccme_med}
\alias{white_hccme_med}
\title{Calculate the Heteroskedasticity-Consistent Covariance Matrix Estimator less slowly.}
\usage{
white_hccme_med(lm_in)
}
\arguments{
\item{lm_in}{is an object of class "\code{lm}" that is the output of a
linear regression model.}
}
\value{
A list containing two elements. The first is the HCCME covariance
matrix. the second is a data frame of the form of \code{coef(summary(lm_in))},
except with standard errors and the remaining statistics adjusted
for heteroskedasticity.
}
\description{
\code{white_hccme_med} calculates the HCCME for a linear regression model,
to match the calculation for un-aggregated data, i.e. data
with weights either \code{NULL} or equal to 1.
It calculates the covariance matrix from first principles, i.e. with loops,
but with loops in the rows and columns of the covariance matrix, not in the sample size.
}
\seealso{
\code{white_hccme}, which uses \code{vcovHC} function in the
\code{sandwich} package.
}
