% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregress.R
\name{adj_wtd_lm_summary}
\alias{adj_wtd_lm_summary}
\title{Create a revised summary of an lm object.}
\usage{
adj_wtd_lm_summary(wtd_lm)
}
\arguments{
\item{wtd_lm}{an lm object from a regression with frequencies as the weights.}
}
\value{
The revised summary of the lm object \code{wtd_lm}.
}
\description{
Create a revised summary of an lm object.
}
